requirements.txt
Django>=3.2
psycopg2-binary>=2.8
djangorestframework>=3.9.2
django-cors-headers>=3.3.0
